requirements/dev.txt
-r prod.txt

# Developer Tools
ipdb==0.13.8
black==21.9b0
django-extensions==3.1.3
pre-commit==2.15.0

# Testing
factory-boy==3.2.0
django-nose==1.4.7
nose-progressive==1.5.2
coverage==5.5

# Thirdparty
django-phone-field==1.8.1
